@@ -1,5 +1,5 @@
 module aoc
 
-go 1.17
+go 1.19
 
 require golang.org/x/net v0.0.0-20211205041911-012df41ee64c
